feat(admin): Hiding questions happen automatically between rounds
- automatically hide questions between rounds. Move the button down near
  the main controls. I believe this is closer UX wise to what our users
  would want either hosting in person or online.
